requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Werkzeug==2.3.7
Pillow==10.0.0
gunicorn==21.2.0
